@@ -12,7 +12,7 @@
     "devDependencies": {
         "@babel/plugin-syntax-dynamic-import": "^7.8.3",
         "@inertiajs/inertia": "^0.11.0",
-        "@inertiajs/inertia-vue": "^0.8.0",
+        "@inertiajs/inertia-vue3": "^0.6.0",
         "@inertiajs/progress": "^0.2.7",
         "axios": "^0.25",
         "browser-sync": "^2.27.7",
@@ -23,8 +23,7 @@
         "resolve-url-loader": "^5.0.0",
         "sass": "^1.49.8",
         "sass-loader": "^12.1.0",
-        "vue": "2.6.14",
-        "vue-loader": "^15.9.7",
-        "vue-template-compiler": "^2.6.14"
+        "vue": "^3.2.31",
+        "vue-loader": "^16.2.0"
     }
 }
