@@ -1,17 +1,14 @@
-import Vue from 'vue'
-import { createInertiaApp } from '@inertiajs/inertia-vue'
+import { createApp, h } from 'vue'
+import { createInertiaApp } from '@inertiajs/inertia-vue3'
 import { InertiaProgress } from '@inertiajs/progress'
-
-Vue.prototype.$route = route
 
 createInertiaApp({
   resolve: (name) => import(`./pages/${name}`),
   setup({ el, App, props, plugin }) {
-    Vue.use(plugin)
-
-    new Vue({
-      render: (h) => h(App, props),
-    }).$mount(el)
+    createApp({ render: () => h(App, props) })
+      .use(plugin)
+      .mixin({ methods: { route } })
+      .mount(el)
   },
 })
 
